+++ include/clap/ext/draft/ambisonic.h
@@ -4,7 +4,7 @@
 
 // This extension can be used to specify the channel mapping used by the plugin.
 
-static CLAP_CONSTEXPR const char CLAP_EXT_AMBISONIC[] = "clap.ambisonic.draft/2";
+static CLAP_CONSTEXPR const char CLAP_EXT_AMBISONIC[] = "clap.ambisonic.draft/3";
 
 static CLAP_CONSTEXPR const char CLAP_PORT_AMBISONIC[] = "ambisonic";
 
@@ -12,15 +12,15 @@ static CLAP_CONSTEXPR const char CLAP_PORT_AMBISONIC[] = "ambisonic";
 extern "C" {
 #endif
 
-enum {
+enum clap_ambisonic_ordering {
    // FuMa channel ordering
-   CLAP_AMBISONIC_FUMA = 0,
+   CLAP_AMBISONIC_ORDERING_FUMA = 0,
 
    // ACN channel ordering
-   CLAP_AMBISONIC_ACN = 1,
+   CLAP_AMBISONIC_ORDERING_ACN = 1,
 };
 
-enum {
+enum clap_ambisonic_normalization {
    CLAP_AMBISONIC_NORMALIZATION_MAXN = 0,
    CLAP_AMBISONIC_NORMALIZATION_SN3D = 1,
    CLAP_AMBISONIC_NORMALIZATION_N3D = 2,
@@ -28,21 +28,26 @@ enum {
    CLAP_AMBISONIC_NORMALIZATION_N2D = 4,
 };
 
-typedef struct clap_ambisonic_info {
-   uint32_t ordering;
-   uint32_t normalization;
-} clap_ambisonic_info_t;
+typedef struct clap_ambisonic_config {
+   uint32_t ordering;      // see clap_ambisonic_ordering
+   uint32_t normalization; // see clap_ambisonic_normalization
+} clap_ambisonic_config_t;
 
 typedef struct clap_plugin_ambisonic {
+   // Returns true if the given configuration is supported.
+   // [main-thread]
+   bool(CLAP_ABI *is_config_supported)(const clap_plugin_t           *plugin,
+                                       const clap_ambisonic_config_t *config);
+
    // Returns true on success
    //
    // config_id: the configuration id, see clap_plugin_audio_ports_config.
    // If config_id is CLAP_INVALID_ID, then this function queries the current port info.
    // [main-thread]
-   bool(CLAP_ABI *get_info)(const clap_plugin_t   *plugin,
-                            bool                   is_input,
-                            uint32_t               port_index,
-                            clap_ambisonic_info_t *info);
+   bool(CLAP_ABI *get_config)(const clap_plugin_t     *plugin,
+                              bool                     is_input,
+                              uint32_t                 port_index,
+                              clap_ambisonic_config_t *config);
 
 } clap_plugin_ambisonic_t;